refactor: split user_scope_test.rs into 6 modules
- Split 937-line test file into focused modules
- subject_test.rs, request_test.rs, creator_rule_test.rs,
  rule_evaluation_test.rs, validation_test.rs, template_test.rs
- All 39 tests pass
- Each file under 500 lines (max 208 lines)

Source: issue #214
Type: refactor
