+++ src/permission/tests/mod.rs
@@ -1,2 +1,7 @@
 //! Permission module tests
-mod user_scope_test;
+mod creator_rule_test;
+mod request_test;
+mod rule_evaluation_test;
+mod subject_test;
+mod template_test;
+mod validation_test;
